style: split solve logic to features layer
