+++ FrontEnd/src/widgets/siblingRivalry/ui/SolveWidget.tsx
@@ -1,15 +1,8 @@
+import { useSolveGame } from "@/features/solve-game";
 import { Button } from "@chakra-ui/react";
-import { useCallback, useState } from "react";
 
 export const SolveWidget = () => {
-  // TODO: mocking loading state
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const handleSolveClick = useCallback(() => {
-    setIsLoading(true);
-    setTimeout(() => {
-      setIsLoading(false);
-    }, 2000);
-  }, []);
+  const { isLoading, handleSolveClick } = useSolveGame();
   return (
     <Button
       size="lg"
